requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy==2.0.35
alembic==1.13.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
python-dotenv==1.0.0